### 2020-08-14 #2570 @whedon
[PRE REVIEW]: Minerva: a light-weight, narrative image browser for multiplexed tissue images
**Submitting author:** @thejohnhoffer (<a href="http://orcid.org/0000-0002-5877-4266">John Hoffer</a>)
**Repository:** <a href="https://github.com/labsyspharm/minerva-story" target ="_blank">https://github.com/labsyspharm/minerva-story</a>
**Version:** v1.0.0
