### 2020-08-14 #2570 @kyleniemeyer
Hi @thejohnhoffer, thanks for your submission to JOSS. It looks like your article is missing one of the required elements: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
