### 2020-08-17 #2570 @will-moore
@jni I'd be happy to review this.
Not reviewed for JOSS before, so might need some pointers...
