### 2020-08-17 #2570 @jni
Awesome, thank you so much, @will-moore!

This is actually my second review as editor, but thankfully the JOSS process well documented, essentially:
- read the reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
- read the review criteria: https://joss.readthedocs.io/en/latest/review_criteria.html
- work your way through the checklist, which will also appear in the review issue: https://joss.readthedocs.io/en/latest/review_checklist.html

All this will happen *after* I ask whedon to start the review (shortly), which will create a separate REVIEW issue, on which you'll be pinged.

Again, thank you! :tada:
