### 2020-08-17 #2570 @jni
@sofroniewn, thanks for agreeing to review, and please see my [comment above](https://github.com/openjournals/joss-reviews/issues/2570#issuecomment-674845258) for a summary of the JOSS review process!
