### 2020-08-17 #2579 @whedon
[REVIEW]: Minerva: a light-weight, narrative image browser for multiplexed tissue images
**Submitting author:** @thejohnhoffer (<a href="http://orcid.org/0000-0002-5877-4266">John Hoffer</a>)
**Repository:** <a href="https://github.com/labsyspharm/minerva-story" target ="_blank">https://github.com/labsyspharm/minerva-story</a>
**Version:** 1.0.0
