### 2020-08-18 #2579 @will-moore
I'm failing at "Is the source code for this software available at the repository url?" for 2 reasons:
 - Seems that all the JS code is minified in https://github.com/labsyspharm/minerva-story/blob/master/_includes/bundle.js and the MinervaStory source code isn't included in the repository
 - Also, half the paper is about https://github.com/labsyspharm/minerva-author so should the checklist above be duplicated for that repo?
