### 2020-08-19 #2579 @jni
I think this means @thejohnhoffer needs to update the code in the repo to be readable, and include the code for minerva-author somewhere in the submission repo as well, but I hope @lpantano can confirm this or clarify...
