### 2020-08-19 #2579 @jni
Just got confirmation from @arfon:

- The authors should unminify all code for review, though that bundle might be external libraries?

- The source code under review should preferably be together in this repo, but otherwise, the authors should point to the source code to be reviewed, *and* they will have to produce a complete bundle for the Zenodo archive at the time of acceptance.
