### 2020-08-24 #2579 @thejohnhoffer
Hi @jni and @will-moore .

I'm sorry for the confusion. We have indeed split the code for this project among four separate repositories. Instructions for where to access these repositories is hidden on the [Minerva Story Wiki](https://github.com/labsyspharm/minerva-story/wiki/For-Developers), but I'll replicate that here.

### Minerva Story
The GitHub Pages site build is stored at [minerva-story](https://github.com/labsyspharm/minerva-story). The source code for the minified bundle is stored at [minerva-browser](https://github.com/labsyspharm/minerva-browser).

### Minerva Author
The Python Flask server along with automated testing is stored at [minerva-author](https://github.com/labsyspharm/minerva-author). The React UI is stored at [minerva-author-ui](https://github.com/labsyspharm/minerva-author-ui)

Again, I do apologize. Minerva Story is the repository to host the final product in the form for a Jekyll blog, but Minerva as a whole contains much more informative code in the other three repositories.
