### 2020-08-24 #2579 @will-moore
Hi @thejohnhoffer, thanks for that.
Is there any more info on how to build the source code from https://github.com/labsyspharm/minerva-browser into the bundle.js at https://github.com/labsyspharm/minerva-story/blob/master/_includes/bundle.js?
I can see a webpack build config etc for minerva-author-ui but not for minerva-browser?
 Many thanks,
  Will.
