### 2020-08-25 #2579 @thejohnhoffer
Hi @will-moore ,

We have a webpack build config, but it has not yet been integrated into either minerva-story or minerva-browser. Today, I will refactor the code slightly to include in the minerva-story repo a proper build setup for the minerva-browser bundle.
