### 2020-08-25 22f04f4
add webpack build scripts, remove node modules
