+++ index.js
@@ -1,0 +1,6 @@
+import mb from 'minerva-browser'
+
+export default {
+  build_page: mb.build_page,
+  get_links_alias: mb.get_links_alias
+}
