+++ scripts.js
@@ -1,0 +1,3 @@
+window.$ = window.jQuery = require('jquery');
+window.Popper = require('popper.js').default;
+require('bootstrap/dist/js/bootstrap');
