### 2020-08-25 3e7b5ed
allow jekyll to reference bootstrap scss from node_modules
