+++ _config.yml
@@ -4,9 +4,8 @@ description: A Simple Approach to Highly Multiplexed Cell and Tissue Imaging
 sass:
     load_paths:
         - _scss
-        - node_modules
+        - node_modules/bootstrap
 plugins:
     - jekyll-redirect-from
-exclude: []
 google_analytics: UA-41775478-2
 repository: labsyspharm/minerva-story
