### 2020-08-25 #2579 @thejohnhoffer
Hi @will-moore,

I've updated Minerva Story [to include webpack instructions](https://github.com/labsyspharm/minerva-story#rebuilding-minerva-browser) for rebuilding Minerva Browser.
