### 2020-09-01 #2579 @will-moore
I've checked off all the items that I feel able to just now.
I have issues setting up ruby on my Mac to test the install instructions, but I'll continue to look at that.
Other issues are:
 - I'm not sure if API documentation is essential, given that this is presented as a GUI tool rather than a software library?
 - I couldn't see any Automated tests or community guidelines for contributing.
 - The paper doesn't compare other tools for viewing large multiplexed images on the web.
