### 2020-09-07 #2579 @jni
Just a status report from my perspective:

- @thejohnhoffer, could you address the [comments](https://github.com/openjournals/joss-reviews/issues/2579#issuecomment-684823383) from @will-moore above?
- We are still waiting for review from @sofroniewn, but he has already noted that it's on the way.
