+++ paper/paper.bib
@@ -5,8 +5,7 @@
   	journal = {Cell Systems},
 	doi = {10.1016/j.cels.2016.03.008},
   	year = 2016,
-  	month = mar,
-	adsurl = {https://linkinghub.elsevier.com/retrieve/pii/S2405471216301065}
+  	month = mar
 }
 
 @article{Rozenblatt-Rosen:2020,
@@ -15,8 +14,7 @@
 	journal = {Cell},
 	doi = {10.1016/j.cell.2020.03.053},
 	year = 2020,
-	month = apr,
-	adsurl = {https://www-sciencedirect-com.ezp-prod1.hul.harvard.edu/science/article/pii/S0092867420303469}
+	month = apr
 }
 
 @article{Rashid:2020,
@@ -24,10 +22,9 @@
 	title = {Interpretative guides for interacting with tissue atlas and digital pathology data using the Minerva browser},
 	journal = {bioRxiv},
 	archiveprefix = {bioRxiv},
-	doi = {10.1051/0004-6361/201322068},
+	doi = {10.1101/2020.03.27.001834},
 	year = 2020,
-	month = mar,
-	adsurl = {https://www.biorxiv.org/content/10.1101/2020.03.27.001834v1}
+	month = mar
 }
 
 @misc{OpenSeadragon,
@@ -50,6 +47,5 @@
   	journal = {Cancer Discovery},
 	doi = {10.1158/2159-8290.CD-12-0095},
   	year = 2012,
-  	month = may,
-	adsurl = {https://cancerdiscovery.aacrjournals.org/content/2/5/401.long}
+  	month = may
 }
