### 2020-09-09 #2579 @jmuhlich
Regarding the comparison to other tools, we thought the reference to our other publication Rashid 2020 (in review at another journal) would be sufficient -- that work surveys the field much more comprehensively than we have room for here. Would you prefer we include a small summary of that review in this manuscript? If not, we can at least make it more clear in the text that this information can be found in that reference.

We'll certainly add community guidelines and think about what we can create automated tests for.
