### 2020-09-11 #2579 @will-moore
Hi @jmuhlich, Thanks, I see that Rashid et.al exactly ticks the box of "describe how this software compares to other commonly-used packages" but I guess that could be made a little clearer.
I found that the DOI for Rashid seems to be incorrect: https://doi.org/10.1051/0004-6361/201322068
Google led me to https://www.biorxiv.org/content/10.1101/2020.03.27.001834v1 but I ended up reading the current version at https://www.biorxiv.org/content/10.1101/2020.03.27.001834v2.
Thanks,
