+++ webpack.config.js
@@ -4,7 +4,7 @@ IndexLoader = {
   entry: './index.js',
   output: {
     filename: 'bundle.js',
-    path: path.resolve(__dirname, '_includes'),
+    path: path.resolve(__dirname, 'includes'),
     libraryTarget: 'var',
     library: 'MinervaStory'
   }
@@ -14,7 +14,7 @@ ScriptsLoader = {
   entry: './scripts.js',
   output: {
     filename: 'scripts.js',
-    path: path.resolve(__dirname, '_includes')
+    path: path.resolve(__dirname, 'includes')
   }
 }
 
