### 2020-09-16 0c90e29
remove all jekyll dependencies
