+++ index.md
@@ -1,5 +1,0 @@
----
-layout: osd-exhibit
-paper: example
-figure: example
----
