+++ webpack.config.js
@@ -7,6 +7,21 @@ IndexLoader = {
     path: path.resolve(__dirname, 'includes'),
     libraryTarget: 'var',
     library: 'MinervaStory'
+  },
+  module: {
+    rules: [
+      {
+        test: /\.s[ac]ss$/i,
+        use: [
+          // Creates `style` nodes from JS strings
+          'style-loader',
+          // Translates CSS into CommonJS
+          'css-loader',
+          // Compiles Sass to CSS
+          'sass-loader',
+        ],
+      },
+    ]
   }
 }
 
