+++ paper/paper.md
@@ -55,7 +55,7 @@ bibliography: paper.bib
 
 # Summary
 
-Advances in highly multiplexed tissue imaging are transforming our understanding of human biology by enabling detection and localization of 10-100 proteins at subcellular resolution [@Bodenmiller:2016]. Efforts are now underway to create public atlases of multiplexed images of normal and diseased tissues [@Rozenblatt-Rosen:2020]. Both research and clinical applications of tissue imaging benefit from recording data from complete specimens so that data on cell state and composition can be studied in the context of overall tissue architecture. As a practical matter, specimen size is limited by the dimensions of microscopy slides (2.5 x 7.5 cm or ~2-8 cm$^2$ of tissue depending on shape). With current microscopy technology, specimens of this size can be imaged at sub-micron resolution across ~60 spectral channels and ~10$^6$ cells, resulting in image files of terabyte size. However, the rich detail and multi-scale properties of these images pose a substantial computational challenge [@Rashid:2020].
+Advances in highly multiplexed tissue imaging are transforming our understanding of human biology by enabling detection and localization of 10-100 proteins at subcellular resolution [@Bodenmiller:2016]. Efforts are now underway to create public atlases of multiplexed images of normal and diseased tissues [@Rozenblatt-Rosen:2020]. Both research and clinical applications of tissue imaging benefit from recording data from complete specimens so that data on cell state and composition can be studied in the context of overall tissue architecture. As a practical matter, specimen size is limited by the dimensions of microscopy slides (2.5 x 7.5 cm or ~2-8 cm$^2$ of tissue depending on shape). With current microscopy technology, specimens of this size can be imaged at sub-micron resolution across ~60 spectral channels and ~10$^6$ cells, resulting in image files of terabyte size. However, the rich detail and multi-scale properties of these images pose a substantial computational challenge [@Rashid:2020]. See [@Rashid:2020] for an comparison of existing visualization tools targetting these multiplexed tissue images.
 
 In this paper we describe a new open-source visualization tool, Minerva, which facilitates intuitive real-time exploration of large multiplexed images on the web. Minerva employs the OpenSeadragon [@OpenSeadragon] framework to render images at multiple resolutions and makes it possible to pan and zoom across across images in a process analogous to Google Maps. However, tissues contain many specialized structures recognizable by pathologists and histologists but not necessarily by many other scientific or medical users. To capitalize on specialized histology expertise we require software that mimics the current practice in which a pathologists sits alongside a colleague and reviews a specimen by moving from point to point and switching between high and low magnifications.
 
