### 2020-09-18 #2579 @thejohnhoffer
I've updated the paper to mention that Rashid 2020 contains a comparison of existing software, I've also updated the DOI for that particular paper. We've also added [contribution guidelines to the wiki](https://github.com/labsyspharm/minerva-story/wiki/Contribution-Guidelines). Also, I'd like to point out that we have tests for Minerva Author [here](https://github.com/labsyspharm/minerva-author/blob/master/src/test_app.py) and [here](https://github.com/labsyspharm/minerva-author/blob/master/src/test_render_jpg.py) that can be run with a `pytest` command.

@will-moore are you still having trouble install Ruby on MacOS? On our wiki, we have two options for installing [Ruby on MacOS](https://github.com/labsyspharm/minerva-story/wiki/How-to-run-Minerva-Story-locally%3F#installing-ruby-on-macos). The first option uses RVM, which can avoid a lot of the issues on MacOS with the system version of Ruby. If you're still stuck, would you be able to share the error you're seeing?
