### 2020-09-23 07cbad2
move some dependencies to minerva-browser
