### 2020-09-23 aba48f1
add marker and cell type links table to minerva-browser
