+++ index.js
@@ -10,6 +10,5 @@ window.AuthenticationDetails = AmazonCognitoIdentity.AuthenticationDetails;
 import mb from 'minerva-browser'
 
 export default {
-  build_page: mb.build_page,
-  get_links_alias: mb.get_links_alias
+  build_page: mb.build_page
 }
