### 2020-09-23 49ddc56
move openseadragon and plugins to minerva-browser
