### 2020-09-23 f4de108
move more dependencies to minerva-browser
