### 2020-09-24 #2579 @will-moore
Thanks @thejohnhoffer. The install for minerva-story and minerva-author worked fine for me. The only step missing from https://github.com/labsyspharm/minerva-author#installing that I had to do was `brew install openslide`.
I was able to run the example from minerva-story.
For minerva-author I can start the app but currently I never get further than `Opening file: /path/to/image.tiff`.
This seems to hang for me.
But don't think that needs to block any of my checklist, so I think I'm done.
