### 2020-09-27 labsyspharm/minerva-story#63 @sofroniewn
minor installation process updates
hello! I'm one of the JOSS reviewers for your submission here https://github.com/openjournals/joss-reviews/issues/2579, just getting going with things.

I was able to install successfully, but might suggest adding a few extra clarifying lines where I had to "guess" what to do.

First was I ran into a permissions error on `gem install jekyll bundler`. I used `sudo gem install jekyll bundler` and it worked fine, but am not sure if that is recommended. A comment to that effect might help a beginner.

Before running `bundle install` I had to clone the repository and navigate to the main directory. While this may be obvious in retrospect a line or two explaining that could also help a beginner (the next instruction contains the helpful advice `In the project folder, run command`)
