### 2020-09-27 #2579 @sofroniewn
Thanks @jni, permissions updated, I've begun the review process, including making some minor suggestions here https://github.com/labsyspharm/minerva-story/issues/63. Will press on with further review promptly
