### 2020-09-28 #2579 @sofroniewn
I've gone through the software paper and am happy with it, I think the added information that Rashid 2020 contains a comparison of existing software is sufficient to detail the state of the field.

I will continue to review the documentation and the functionality.
