### 2020-09-29 labsyspharm/minerva-story#63 @jmuhlich
Hi Nick, thanks for the feedback. We can definitely add these to the docs. Since the initial manuscript submission we've been working on removing Jekyll and thus Ruby/Gem/Bundler from the workflow entirely, instead adding live preview support directly to Minerva Author. Not sure it's quite ready to replace the current Jekyll build process and revise the manuscript though. We can discuss in our daily meeting tomorrow.
