### 2020-10-04 labsyspharm/minerva-story#64 @sofroniewn
Link to docs in readme
As part of the https://github.com/openjournals/joss-reviews/issues/2579, I think it would be nice if there were more clear links to the documentation in the readme. It took me a while to find this link https://github.com/labsyspharm/minerva-story/wiki in the `About` section (maybe i'm just not so used to the new github site layout). People would also find this link https://www.cycif.org/software/minerva helpful that is referenced at the bottom of the JOSS manuscript.
