### 2020-10-04 labsyspharm/minerva-story#65 @sofroniewn
Add explanation to readme of repo organization
As part of my review for https://github.com/openjournals/joss-reviews/issues/2579, I think this readme could also contain an additional explanation of the repository organization - maybe something along the lines of this comment https://github.com/openjournals/joss-reviews/issues/2579#issuecomment-679235692 and a reference to `minerva-browser` in particular https://github.com/labsyspharm/minerva-browser. You could in fact write a generic portion of the readme that could be reused on all the related repos indicating their interconnectedness. This could help a new contributor get oriented in the project.
