### 2020-10-04 #2579 @sofroniewn
I made an issue asking for the documentation links to be added to the readme so they are easier to discover.

I've gone through the documentation and functionality and found them to be satisfactory.

I found [contributing guidelines](https://github.com/labsyspharm/minerva-story/wiki/Contribution-Guidelines) here.

I wasn't able to find a full API reference say for code in https://github.com/labsyspharm/minerva-browser, but I think given the nature of this project as a web app, I don't think the target audience or potential contributors are expecting a full API reference and the code does contain helpful comments and appears clearly written, so I think it is sufficiently documented.

I wasn't able to find any continuous integration or automated tests, so I have left that checkbox blank, but I think that is not unusual for web projects of this nature and so I don't have any additional concerns.

I really like the examples present in the `featured stories` on [the main software website](https://www.cycif.org/software/minerva). They are a great way to get a sense for what can be built with the tool and a highlight of the submission.

I have not made my own story, but I was able to install and successfully install and launch the software and am convinced by the high quality examples (which I have run in my browser) that the functionality exists as described in the manuscript and the performance of the software (interactive panning/ zooming of large datasets) matches the expectations described in the manuscript.

I now consider my review complete and approve this submission, but I'm happy to expand on any of my statements or look into things more at the request of @jni.

Great job authors, congrats on a really nice piece of software and manuscript. I'm excited to see the community pick this up!!
