### 2020-10-05 #2579 @jni
Thank you @will-moore and @sofroniewn for your reviews!

@will-moore can you confirm that you're happy to accept the manuscript as it stands?

Thanks everyone for the interaction to improve the paper and the software thus far!
