### 2020-10-05 labsyspharm/minerva-story#63 @thejohnhoffer
Hi @sofroniewn,

We've since updated Minerva Story to no longer require Ruby or Jekyll. I've also updated the README.md of this repository to be explicit about cloning the repository and changing to the required directory. We've included installation instructions for installing `npm` and `npx` to allow the users to run a simple `http-server`, though any simple HTTP server would work just as well.

We'll soon update the manuscript to reflect these changes.
