+++ README.md
@@ -1,5 +1,15 @@
 # Minerva Story 
 
+## Minerva Repository Structure
+
+### Minerva Story
+The GitHub Pages site build is stored at [minerva-story](https://github.com/labsyspharm/minerva-story). The source code for the minified bundle is stored at [minerva-browser](https://github.com/labsyspharm/minerva-browser).
+
+### Minerva Author
+The Python Flask server along with automated testing is stored at [minerva-author](https://github.com/labsyspharm/minerva-author). The React UI is stored at [minerva-author-ui](https://github.com/labsyspharm/minerva-author-ui)
+
+## Running Minerva Story
+
 ### Clone repository
 
 ```bash
