### 2020-10-05 labsyspharm/minerva-story#65 @thejohnhoffer
Hi @sofroniewn,

I've added the snippet from that review comment to all four of the Minerva README.md files.

Thanks,
John
