### 2020-10-05 labsyspharm/minerva-story#64 @thejohnhoffer
Hi @sofroniewn,

I've added both links to the Minerva Story README.md.

Thanks,
John
