+++ paper/paper.md
@@ -65,9 +65,9 @@ Minerva is designed to generate precisely these types of interactive guides or "
 
 Minerva is not designed to solve all analytical and visualization problems encountered in multiplexed tissue imaging. Instead, it is a publication tool specialized to the task of making data shareable and broadly intelligible without requiring specialized software on the user side. As such, Minerva is designed to be one component in an ecosystem of interoperable, open-source software tools.
 
-Minerva comprises two components, _Minerva Story_ and _Minerva Author_. _Minerva Story_ is a single-page web application for presenting an image and its narrative to end users. OpenSeadragon[@OpenSeadragon] is used to render tiled JPEG image pyramids overlaid with the author's narrative text, graphical annotations and data plots. Audio presentation of the narrative text is optionally provided through integration with Amazon's Polly text-to-speech service. The image pyramid URL and all narrative details are loaded from an independent YAML story definition file. Minerva Story is deployed via the Jekyll [@Jekyll] static site generator which allows users to integrate multiple stories and accompanying content into a simple website and host it through GitHub Pages or any other web host supporting static content such as Amazon S3.
+Minerva comprises two components, _Minerva Story_ and _Minerva Author_. _Minerva Story_ is a single-page web application for presenting an image and its narrative to end users. OpenSeadragon[@OpenSeadragon] is used to render tiled JPEG image pyramids overlaid with the author's narrative text, graphical annotations and data plots. Audio presentation of the narrative text is optionally provided through integration with Amazon's Polly text-to-speech service. The image pyramid URL and all narrative details are loaded from an independent JSON story definition file. Minerva Story can be hosted through GitHub Pages or any other web host supporting static content such as Amazon S3.
 
-_Minerva Author_ is a desktop application for constructing narratives (stories) for Minerva Story. It is a JavaScript React web application with a Python Flask backend, packaged with PyInstaller as a native application. To create a narrative, the author first imports an image in standard OME-TIFF pyramid or SVS formats. Both RGB images (brightfield, H&E, immunohistochemistry, etc.) and multi-channel fluorescence images (immunofluorescence, CODEX, CyCIF, etc.) are supported. Fluorescence image rendering can be controlled through per-channel contrast adjustment and pseudocoloring. The author then adds one or more story waypoints. For each waypoint, the author can type a text description, draw polygon or arrow annotations to highlight specific cells, regions, and histological features, and choose specific image channels to present. See \autoref{fig:author} for the Minerva Author interface. After the waypoints are complete, Minerva Author renders the input image into one or more RGB JPEG image pyramids and produces a YAML story definition file. Finally, an author  uploads the JPEG images (vastly smaller than the raw image data) and YAML file to a web host along with the HTML files produced by Jekyll to generate a finished, browsable Minerva Story. Story waypoints can be also augmented with interactive linked data visualizations. Adding data visualization currently requires manually editing the YAML file but a version of Minerva Author is in development for adding these types of visualizations natively.
[…]
