+++ paper/paper.bib
@@ -34,13 +34,6 @@
 	url = {https://github.com/openseadragon/openseadragon}
 }
 
-@misc{Jekyll,
-	title = {Jekyll},
-	publisher = {GitHub},
-	journal = {GitHub repository},
-	url = {https://github.com/jekyll/jekyll}
-}
-
 @article{Cerami:2012,
 	author = {{Cerami}, E. and {Gao}, J. and {Dogrusoz}, U. and {Gross}, B.~E. and {Sumer}, S.~O. and {Aksoy}, B.~A. and {Jacobsen}, A. and {Byrne}, C.~J. and {Heuer}, M.~L. and {Larsson}, E. and {Antipin}, Y. and {Reva}, B. and {Goldberg}, A.~P and {Sander}, C. and {Schultz}, N.},
   	title = {The cBio cancer genomics portal: an open platform for exploring multidimensional cancer genomics data},
