### 2020-10-07 #2579 @thejohnhoffer
I'd like to mention that to resolve [issues with ruby dependencies](https://github.com/labsyspharm/minerva-story/issues/63), we have updated the manuscript and the Minerva Story repository to no longer require the use of Jekyll (or Ruby).
