### 2020-10-12 #2579 @jni
LOL This is my first time accepting a JOSS paper so bear with me... 😅
