### 2020-10-12 #2579 @jni
@thejohnhoffer @jmuhlich I checked the final proof for typos/small errors, and found just two very minor issues:

- the title for Rashid R et al is "Interpretative guides for interacting with tissue atlas and digital pathology data using the minerva browser" in the paper, but "Online narrative guides for illuminating tissue atlas data and digital pathology images" on biorxiv.
- "targetting" should be changed to "targeting" (according to [this page](https://english.stackexchange.com/questions/205815/is-targetted-a-standard-british-english-spelling)).

After those small issues are fixed, we need you to make a tagged release and archive of the software, and to report the version number and DOI here. (That was Whedon's issue above 😂.)
