+++ paper/paper.bib
@@ -19,7 +19,7 @@
 
 @article{Rashid:2020,
 	author = {{Rashid}, R. and {Chen}, Y.~A. and {Hoffer}, J. and {Muhlich}, J.~L. and {Chen}, Y.~A. and {Lin}, J.~R. and {Krueger}, R. and {Pfister}, H. and {Mitchell}, R. and {Santagata}, S. and {Sorger}, P.~K.},
-	title = {Interpretative guides for interacting with tissue atlas and digital pathology data using the Minerva browser},
+	title = {Online narrative guides for illuminating tissue atlas data and digital pathology images},
 	journal = {bioRxiv},
 	archiveprefix = {bioRxiv},
 	doi = {10.1101/2020.03.27.001834},
