### 2020-10-13 #2579 @thejohnhoffer
@jni I've fixed those two issues with the paper. Thanks for finding them. For the release and archive of our software, do you have any suggestions as to how we should handle our four separately versioned repositories? Also, for the archive process, my current idea is to [follow GitHub's guide for creating citable archives.](https://guides.github.com/activities/citable-code/)
