### 2020-10-13 #2579 @jni
@thejohnhoffer great questions! To me it sounds good but I hope @lpantano and/or @arfon can provide a more definitive answer.
