### 2020-10-14 #2579 @lpantano
I think that sounds good, I will let @arfon to re-confirm. As far as there is an isolated download from Zenodo that makes this reproducible to this point, it should be good (i think).
