### 2020-10-14 #2579 @arfon
> For the archive process, our current idea is to use Zenodo to archive a zip file containing exports of our four separate repositories. We would version releases for our repositories separately, and we'd have a README in the Zenodo archive pointing to the four separate releases on github. The paper itself would be within the minerva-story folder of the Zenodo archive. Would this be an okay way to handle an archive for our four repositories?

Sounds great. Thanks for confirming @thejohnhoffer.
