### 2020-10-15 315e86d
2 minor proofing changes in paper (#66)
